+++ CHANGELOG.md
@@ -15,7 +15,7 @@ New functionality added in a backwards-compatible manner:
 * [OLMIS-6358](https://openlmis.atlassian.net/browse/OLMIS-6358): Added versioning to Facility Type Approved Product resource.
 * [OLMIS-6430](https://openlmis.atlassian.net/browse/OLMIS-6430): Added ability to search orderables by identities.
   * To use this feature please use the new `POST /api/orderables/search` endpoint
-  * For now, both `id` and `versionId` fields must be provided in the request body to find an orderable
+  * For now, both `id` and `versionNumber` fields must be provided in the request body to find an orderable
 
 Bug fixes:
 * [OLMIS-6188](https://openlmis.atlassian.net/browse/OLMIS-6188): Added validation for name when creating a processing period.